\name{chronometre-package}
\alias{chronometre-package}
\alias{chronometre}
\docType{package}
\title{\packageTitle{chronometre}}
\description{\packageDescription{chronometre}}
\section{Package Content}{\packageIndices{chronometre}}
\author{\packageAuthor{chronometre}}
\section{Maintainer}{\packageMaintainer{chronometre}}
\keyword{package}
